backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0